all functionality with middlewares too donw

+++ src/controllers/category.controller.js
@@ -25,8 +25,15 @@ const getCategory = async (req, res) => {
 
 const createCategory = async (req, res) => {
   try {
+    if (req.user.role != "admin") {
+      return res.status(500).send({
+        message: "only admin can add category",
+      });
+    }
     const category = await Category.create(req.body);
-    return res.status(201).send({ message: "Category created successfully" });
+    return res
+      .status(201)
+      .send({ message: "Category created successfully", data: category });
   } catch (error) {
     return res
       .status(500)
@@ -37,6 +44,11 @@ const createCategory = async (req, res) => {
 const updateCategory = async (req, res) => {
   const { id } = req.params;
   try {
+    if (req.user.role != "admin") {
+      return res.status(500).send({
+        message: "only admin can add category",
+      });
+    }
     const category = await Category.findByIdAndUpdate(id, req.body, {
       new: true,
     });
@@ -51,6 +63,11 @@ const updateCategory = async (req, res) => {
 const deleteCategory = async (req, res) => {
   const { id } = req.params;
   try {
+    if (req.user.role != "admin") {
+      return res.status(500).send({
+        message: "only admin can add category",
+      });
+    }
     const category = await Category.findByIdAndDelete(id);
     return res.status(200).send({ message: "Category deleted successfully" });
   } catch (error) {
